Rewrite hard to understand QuestionGameUploadForm::show()
The direct parent is the UploadForm class, which calls HTMLForm::show()
and UploadForm::addUploadJS(). In other words, the reason for the
override was to disable the addUploadJS() call. This patch suggests
another solution that does the same.

This also removes the redundant trySubmit() which does the exact same
as the parent.

Change-Id: I2ee8af759adb5e01aee43ff5dc801686e790ccdd