Unbreak hacky inline JavaScript
Prior to this patch, visiting a quiz page (Special:QuizGameHome, either without or
with the ?questionGameAction=renderPermalink&permalinkID=<numerical ID> URL params)
would be broken, you'd see the text "Loading..." but the quiz question and the answer
options would never load.

Lightly tested and seems to work, but emphasis on the word "lightly".

Change-Id: I5b5d589c4f2aa7e912e460d5b0ff5b071a0471f9