Use the COALESCE() SQL function to cast NULL into 0 when the subquery returns no rows
This can happen if the user in question hasn't answered any quizzes correctly, for example.
Unlike the SQL COUNT() function, SUM() can and sometimes *will* return NULL.
We don't want NULL, we want a proper number, although the user_stats table columns aren't yet enforced as non-NULLable
(but they should be, but that's a separate issue).

Change-Id: I43e1a3d80b0690f3e4c5adb821d9e88458546b6d